extract switch lane logic from checkout-version.ts to switch-lanes.ts
